Adapted cmd tests
[skip ci]
