--- src/fastoad/cmd/tests/data/cmd_sellar_example/disc1.py
@@ -1,35 +1,0 @@
-"""Sellar discipline 1"""
-#  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
-#  FAST is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
-#  This program is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#  You should have received a copy of the GNU General Public License
-#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-from fastoad.module_management.constants import ModelDomain
-from fastoad.module_management.service_registry import RegisterOpenMDAOSystem
-from .disc1_base import Disc1Base
-
-
-@RegisterOpenMDAOSystem("cmd_test.sellar.disc1", domain=ModelDomain.OTHER, desc="some text")
-class Disc1(Disc1Base):
-    """An OpenMDAO component to encapsulate Disc1 discipline"""
-
-    # pylint: disable=invalid-name
-    def compute(self, inputs, outputs, discrete_inputs=None, discrete_outputs=None):
-        """
-        Evaluates the equation
-        y1 = z1**2 + z2 + x1 - 0.2*y2
-        """
-        z1 = inputs["z"][0]
-        z2 = inputs["z"][1]
-        x1 = inputs["x"]
-        y2 = inputs["y2"]
-
-        outputs["y1"] = z1 ** 2 + z2 + x1 - 0.2 * y2
